refactored frontend into components
