+++ src/App.svelte
@@ -1,15 +1,18 @@
 <script lang="ts">
   import { onMount } from "svelte";
-  import { Connection, LAMPORTS_PER_SOL } from "@solana/web3.js";
-  import { Program, Provider, web3 } from "@project-serum/anchor";
-  import { CANDY_MACHINE_PROGRAM } from "./lib/helpers";
-  import { getMetadata } from "./lib/get-metadata";
-  import { mintOneToken } from "./lib/mint";
-  import { awaitTransactionSignatureConfirmation } from "./lib/connection";
+  import { Connection } from "@solana/web3.js";
+  import { Provider, web3 } from "@project-serum/anchor";
   import { fade } from "svelte/transition";
-  import type { CandyMachineAccount } from "./models";
-  import { TOKEN_PROGRAM_ID } from "@solana/spl-token";
-  import confetti from "canvas-confetti";
+  import Button from "./components/Button.svelte";
+  import Header from "./components/CardHeader.svelte";
+
+  import { candyMachineState, userState } from "./lib/store";
+  import {
+    getCandyMachineState,
+    checkWalletConnected,
+    getUserBalance,
+    existsOwnerSPLToken,
+  } from "./lib/state-helpers";
 
   /***********************************/
   // Customise the app by changing the following variables.
@@ -22,12 +25,9 @@
   /***********************************/
 
   const { solana } = window as any;
-  const cluster = import.meta.env.VITE_APP_SOLANA_NETWORK.toString();
   const rpcUrl = import.meta.env.VITE_APP_SOLANA_RPC_HOST.toString();
-  const txTimeout = 30000;
-  const candyMachineId = new web3.PublicKey(
-    import.meta.env.VITE_APP_CANDY_MACHINE_ID.toString()
-  );
+  const candyMachineId = import.meta.env.VITE_APP_CANDY_MACHINE_ID.toString();
+  const candyMachinePublicKey = new web3.PublicKey(candyMachineId);
   const opts = { preflightCommitment: "processed" };
   const connection = new Connection(rpcUrl);
   const provider = new Provider(
@@ -36,148 +36,34 @@
     opts.preflightCommitment as web3.ConfirmOptions
   );
 
-  let walletPublicKey = "";
-  let itemsRedeemed = 0;
-  let itemsAvailable = 0;
-  let candyMachine: CandyMachineAccount;
-  let isMinting = false;
-  let mintSuccessful = false;
-  let userBalance = 0;
   let siteLoading = true;
-  let solanaExplorerLink = "";
-  let isSoldOut = false;
-  let isActive = false;
-  let price = 0;
-  let userWhitelisted = false;
 
-  // These are computed variables
-  $: date = new Date(candyMachine?.state.goLiveDate?.toNumber() * 1000);
-  $: whitelistToken = candyMachine?.state.whitelistMintSettings?.mint;
-  $: whitelistPrice = candyMachine?.state.whitelistMintSettings?.discountPrice;
-
-  async function checkWalletConnected() {
-    try {
-      const response = await solana.connect({ onlyIfTrusted: true });
-      walletPublicKey = response.publicKey.toString();
-      if (response) {
-        await getUserBalance();
-        await existsOwnerSPLToken();
-      }
-    } catch (error) {}
-  }
-
-  async function connectWallet() {
-    try {
-      const response = await solana.connect();
-      walletPublicKey = response.publicKey.toString();
-      if (response) {
-        await getUserBalance();
-        await existsOwnerSPLToken();
-      }
-    } catch (error) {}
-  }
-
-  async function getCandyMachineState() {
-    try {
-      const idl = await Program.fetchIdl(CANDY_MACHINE_PROGRAM, provider);
-      const program = new Program(idl, CANDY_MACHINE_PROGRAM, provider);
-      candyMachine = await getMetadata(program, candyMachineId);
-      ({ itemsAvailable, itemsRedeemed, isSoldOut, price, isActive } =
-        candyMachine.state);
-    } catch (error) {
-      console.error(error);
-    }
-  }
-
-  async function mint() {
-    try {
-      isMinting = true;
-      if (candyMachine?.program && walletPublicKey) {
-        const mint = web3.Keypair.generate();
-        const mintTxId = (
-          await mintOneToken(
-            candyMachine,
-            new web3.PublicKey(walletPublicKey),
-            mint
-          )
-        )[0];
-        let status: any = { err: true };
-        if (mintTxId) {
-          status = await awaitTransactionSignatureConfirmation(
-            mintTxId,
-            txTimeout,
-            connection,
-            "singleGossip",
-            true
-          );
-        }
-        if (!status?.err) {
-          console.log("Success");
-          displaySuccess(mint.publicKey);
-        } else {
-          console.error("An error occurred");
-        }
-      }
-    } catch (error: any) {
-      console.error("An error occurred ", error);
-    } finally {
-      isMinting = false;
-    }
-  }
-
-  async function existsOwnerSPLToken() {
-    const tokenAccounts = await connection.getParsedTokenAccountsByOwner(
-      new web3.PublicKey(walletPublicKey),
-      {
-        programId: TOKEN_PROGRAM_ID,
-      }
-    );
-    for (let index = 0; index < tokenAccounts.value.length; index++) {
-      const tokenAccount = tokenAccounts.value[index];
-      const tokenAmount = tokenAccount.account.data.parsed.info.tokenAmount;
-
-      const mint = tokenAccount.account.data.parsed.info.mint;
-      if (mint === whitelistToken.toString() && tokenAmount.uiAmount > 0) {
-        console.log("Welcome to the whitelist!");
-        userWhitelisted = true;
-      }
-    }
-  }
-
-  function displaySuccess(mintPublicKey: web3.PublicKey) {
-    // It takes time before state change is cascaded across the network
-    // so we update this variables manually upon successful completion
-    itemsRedeemed += 1;
-    mintSuccessful = true;
-    solanaExplorerLink =
-      cluster == "devnet"
-        ? `https://explorer.solana.com/address/${mintPublicKey}?cluster=devnet`
-        : `https://explorer.solana.com/address/${mintPublicKey}`;
-    throwConfetti();
-  }
-
-  async function getUserBalance() {
-    const pk = new web3.PublicKey(walletPublicKey);
-    try {
-      userBalance = await connection.getBalance(pk);
-    } catch (error) {
-      console.error(error);
-    }
-  }
-
-  function throwConfetti(): void {
-    confetti({
-      particleCount: 200,
-      spread: 70,
-      origin: { y: 0.6 },
-    });
-  }
+  $: itemsRedeemed = $candyMachineState?.state.itemsRedeemed;
+  $: itemsAvailable = $candyMachineState?.state.itemsAvailable;
 
   onMount(async () => {
-    await getCandyMachineState();
-    if (solana) {
-      if (solana.isPhantom) await checkWalletConnected();
+    $candyMachineState = await getCandyMachineState(
+      candyMachinePublicKey,
+      provider
+    );
+
+    if (solana?.isPhantom) {
+      $userState.walletPublicKey = await checkWalletConnected(solana);
+      if ($userState.walletPublicKey) {
+        // Get User Balance
+        $userState.userBalance = await getUserBalance(
+          $userState.walletPublicKey,
+          connection
+        );
+        // Check if user is whitelisted (ie. check if they have token)
+        $userState.isWhiteListed = await existsOwnerSPLToken(
+          $userState.walletPublicKey,
+          connection,
+          $candyMachineState.state.whitelistMintSettings.mint
+        );
+      }
     }
+
     siteLoading = false;
   });
 </script>
@@ -203,32 +89,7 @@
       transition:fade
     >
       <!-- Top Bar -->
-      <div class="justify-end flex p-3">
-        {#if userWhitelisted}
-          <div class="flex mr-auto">
-            <img src="/star.svg" alt="" class="w-5 mr-2" />
-            <div class="my-auto text-gray-600 text-sm">Whitelist</div>
-          </div>
-        {/if}
-        <div class="flex">
-          {#if !walletPublicKey}
-            <span class="my-auto mr-2 rounded-full h-2 w-2 bg-gray-500" />
-            <span class=" my-auto text-gray-600 text-sm"> Not Connected </span>
-          {:else}
-            <div class=" flex flex-col">
-              <div class="flex">
-                <span class="my-auto mr-2 rounded-full h-2 w-2 bg-green-500" />
-                <span class=" my-auto text-gray-600 text-sm">
-                  {walletPublicKey.slice(0, 8)}
-                </span>
-              </div>
-              <div class=" text-xs text-gray-600 text-right ">
-                {(userBalance / LAMPORTS_PER_SOL).toFixed(2)} SOL
-              </div>
-            </div>
-          {/if}
-        </div>
-      </div>
+      <Header />
       <hr />
       <br />
       <!-- Main Body -->
@@ -242,56 +103,15 @@
         <div class="text-sm sm:text-md font-semibold pb-5 text-gray-600 ">
           {DESCRTIPTION}
         </div>
-        <div class="flex flex-col">
-          {#if !solana}
-            <button
-              class=" px-3 py-2 rounded-md  bg-sky-600  hover:bg-sky-700 text-white font-bold"
-              on:click={() => window.open("https://phantom.app/", "_blank")}
-              >Get Phantom Wallet</button
-            >
-          {:else if !walletPublicKey}
-            <button
-              class=" px-3 py-2 rounded-md  bg-sky-600  hover:bg-sky-700 text-white font-bold"
-              on:click={connectWallet}>Connect</button
-            >
-          {:else if !isActive && !userWhitelisted}
-            <button
-              class=" px-3 py-2 rounded-md  bg-sky-600  hover:bg-sky-700 text-white font-bold disabled:bg-gray-400 "
-              disabled={true}>Mint live @ {date.toUTCString()}</button
-            >
-          {:else if isSoldOut}
-            <button
-              class=" px-3 py-2 rounded-md  bg-sky-600  hover:bg-sky-700 text-white font-bold "
-              >Sold Out!</button
-            >
-          {:else}
-            <button
-              class=" px-3 py-2 rounded-md  bg-sky-600  hover:bg-sky-700 text-white font-bold disabled:bg-gray-400"
-              disabled={isMinting}
-              on:click={mint}
-            >
-              {#if isMinting}
-                <span>Minting ...</span>
-              {:else if mintSuccessful}
-                <span>Mint succesful! Mint another?</span>
-              {:else}
-                <span
-                  >Mint ({(
-                    (userWhitelisted ? whitelistPrice : price) /
-                    LAMPORTS_PER_SOL
-                  ).toFixed(2)} SOL)</span
-                >
-              {/if}
-            </button>
-          {/if}
-        </div>
+        <Button {solana} {connection} />
+
         <div class=" tracking-widest font-bold text-sm pt-3 text-gray-400">
           {itemsRedeemed}/{itemsAvailable} claimed
         </div>
         <div class="flex flex-col pt-3">
-          {#if solanaExplorerLink}
+          {#if $userState.solanaExplorerLink}
             <a
-              href={solanaExplorerLink}
+              href={$userState.solanaExplorerLink}
               target="_blank"
               class="text-purple-700 font-semibold  p-1"
               >View on Solana Explorer</a
